Add span finish callback for dynamic data
A single callback can be added per span using
`Span::set_finish_callback`. It is called in the Drop handler, right
before the span is frozen and passed to the tracer channel. Children
inherit their parent's callback, provided they are created with
`Span::child`.

This limitation exists because I didn't want to add the callback member
to `SpanContext`, which only contains data that is propagated across
services. (trace state and baggage) It is documented in the respective
function docs.